requirements.txt
Flask
flask_restful
redis
python-dotenv
